edit user, user card code fixes, currUserCard moved into functions dir, App code clean up
