+++ client/src/components/functions/currentUserData.ts
